Added IS_PERSISTANT tag -- for bullets mostly. Wrote a new shader - 'persistanceclear.sfx'. All it does is clear persistant stuff off the screen that reaches below a certain visibility (I was getting 'streak remnants' before this was in place).
Got rid of some warping ability from warpers, this was slowing the game down
(should probably thread this OR have a [register as warpable] function to reduce
iterations on checking. (profiled the program and this was the main performance
culprit).

New ambient audio file (again), added 'boss music' but it's unused ATM.

Stuck post-processing effects in a stack, executed once every main loop for
performance reasons.